+++ src/app/[locale]/(dashboard)/components/Sidebar/Sidebar.tsx
@@ -9,6 +9,10 @@ import { useMediaQuery } from '@mui/material';
 import OpenSidebarButton from '@/components/OpenSidebarButton/OpenSidebarButton';
 import NewBoardLink from '@/components/NewBoardLink/NewBoardLink';
 import { useBoundStore } from '@/providers/StoreProvider';
+import { useShallow } from 'zustand/react/shallow';
+import { Modal, Button } from '@mui/material';
+
+import './sidebar.css';
 
 export default function Sidebar({ children }: { children: React.ReactNode }) {
   const router = useRouter();
@@ -23,6 +27,12 @@ export default function Sidebar({ children }: { children: React.ReactNode }) {
       setBoardLeaveDialogStatus: state.setBoardLeaveDialogStatus,
     }),
   );
+  const { setBoardIsUpToDate } = useBoundStore(
+    useShallow((state) => ({
+      setBoardIsUpToDate: state.setBoardIsUpToDate,
+    })),
+  );
+  const [setPrompt] = useBoundStore(useShallow((state) => [state.setPrompt]));
   const theme: Theme = useTheme();
   const { isSidebarOpen, toogleIsSidebarOpen } = useBoundStore((state) => ({
     isSidebarOpen: state.isSidebarOpen,
@@ -30,6 +40,32 @@ export default function Sidebar({ children }: { children: React.ReactNode }) {
   }));
 
   const isMobile = useMediaQuery(theme.breakpoints.down('md'));
+
+  const closeDialog = () => setBoardLeaveDialogStatus(false);
+
+  const handleNotSave = () => {
+    if (boardLeaveStatus == 'new') {
+      cleanPrompt();
+      router.push('/board');
+      setBoardLeaveStatus('');
+    } else if (boardLeaveStatus == 'edit') {
+      // const { description, rows, columns, colorScheme, shouldUsePictonizer } =
+      //   data.prompt;
+      // setPrompt({
+      //   description,
+      //   rows,
+      //   columns,
+      //   colorScheme,
+      //   shouldUsePictonizer,
+      // });
+      // data.isSavedBoard
+      //   ? router.push(`/board/${data.id}`)
+      //   : router.push('/board');
+      setBoardLeaveStatus('');
+    }
+    setBoardIsUpToDate();
+    closeDialog();
+  };
 
   return (
     <>
@@ -47,6 +83,30 @@ export default function Sidebar({ children }: { children: React.ReactNode }) {
           <>{children}</>
         </Box>
       </Drawer>
+      <Modal
+        open={boardLeaveDialogStatus}
+        onClose={closeDialog}
+        aria-labelledby="modal-title"
+        aria-describedby="modal-description"
+      >
+        <Box sx={sxStyles.dialog}>
+          <h1 id="modal-title" className="text-center">
+            You didn't save the board
+          </h1>
+          <div className="text-center margin-top-20">
+            <Button
+              variant="outlined"
+              sx={sxStyles.button}
+              onClick={handleNotSave}
+            >
+              Don't Save
+            </Button>
+            <Button variant="outlined" onClick={closeDialog}>
+              Close
+            </Button>
+          </div>
+        </Box>
+      </Modal>
     </>
   );
 }
