+++ src/app/[locale]/(dashboard)/components/Sidebar/sidebar.css
