+++ src/components/DataItem/DataItem.tsx
@@ -8,14 +8,10 @@ import { useFormatter } from 'next-intl';
 import { useBoundStore } from '@/providers/StoreProvider';
 import { useRouter } from 'next/navigation';
 import Box from '@mui/material/Box';
-import { Modal, Button } from '@mui/material';
 import { PromptRecord } from '@/commonTypes/Prompt';
 import { useShallow } from 'zustand/react/shallow';
 import InternalLink from '../InternalLink/InternalLink';
 import useIsSmallScreen from '@/hooks/useIsSmallScreen';
-
-import sxStyles from './styles';
-import './sidebar.css';
 
 export type BaseDataItemType = {
   id: string;
@@ -35,26 +31,25 @@ type Props<DataType extends BaseDataItemType> = {
 export default function DataItem<DataType extends BaseDataItemType>({
   data,
   deleteItem: { deleteData },
-}: Props<DataType extends BaseDataItemType>) {
+}: Props<DataType>) {
   const router = useRouter();
   const { description, rows, columns, colorScheme, shouldUsePictonizer } =
     data.prompt;
   const format = useFormatter();
   const isSmallScreen = useIsSmallScreen();
 
-  const [isOutdated] = useBoundStore(useShallow((state) => [state.isOutdated]));
+  const { isOutdated } = useBoundStore(
+    useShallow((state) => ({
+      isOutdated: state.isOutdated,
+    })),
+  );
 
-  const cleanPrompt = useBoundStore((state) => state.cleanPrompt);
-  const { boardLeaveStatus, setBoardLeaveStatus } = useBoundStore((state) => ({
-    boardLeaveStatus: state.boardLeaveStatus,
+  const { setBoardLeaveStatus } = useBoundStore((state) => ({
     setBoardLeaveStatus: state.setBoardLeaveStatus,
   }));
-  const { boardLeaveDialogStatus, setBoardLeaveDialogStatus } = useBoundStore(
-    (state) => ({
-      boardLeaveDialogStatus: state.boardLeaveDialogStatus,
-      setBoardLeaveDialogStatus: state.setBoardLeaveDialogStatus,
-    }),
-  );
+  const { setBoardLeaveDialogStatus } = useBoundStore((state) => ({
+    setBoardLeaveDialogStatus: state.setBoardLeaveDialogStatus,
+  }));
 
   const [setPrompt, isGenerationPending, toogleIsSidebarOpen] = useBoundStore(
     useShallow((state) => [
@@ -63,26 +58,6 @@ export default function DataItem<DataType extends BaseDataItemType>({
       state.toogleIsSidebarOpen,
     ]),
   );
-
-  const closeDialog = () => setBoardLeaveDialogStatus(false);
-
-  const handleNotSave = () => {
-    if (boardLeaveStatus == 'new') {
-      cleanPrompt();
-      router.push('/board');
-      setBoardLeaveStatus('');
-    } else if (boardLeaveStatus == 'edit') {
-      const { description, rows, columns, colorScheme, shouldUsePictonizer } =
-        data.prompt;
-      setPrompt({description, rows, columns, colorScheme, shouldUsePictonizer});
-      data.isSavedBoard
-        ? router.push(`/board/${data.id}`)
-        : router.push('/board');
-      console.log(prompt);
-      setBoardLeaveStatus('');
-    }
-    closeDialog();
-  };
 
   const onEdit = () => {
     if (isSmallScreen) {
@@ -102,7 +77,7 @@ export default function DataItem<DataType extends BaseDataItemType>({
         : router.push('/board');
     } else {
       setBoardLeaveDialogStatus(true);
-      setBoardLeaveStatus(data);
+      setBoardLeaveStatus('edit');
     }
   };
   return (
@@ -146,30 +121,6 @@ export default function DataItem<DataType extends BaseDataItemType>({
           />
         </Tooltip>
       </ListItem>
-      <Modal
-        open={boardLeaveDialogStatus}
-        onClose={closeDialog}
-        aria-labelledby="modal-title"
-        aria-describedby="modal-description"
-      >
-        <Box sx={sxStyles.dialog}>
-          <h1 id="modal-title" className="text-center">
-            You didn't save the board
-          </h1>
-          <div className="text-center margin-top-20">
-            <Button
-              variant="outlined"
-              sx={sxStyles.button}
-              onClick={handleNotSave}
-            >
-              Don't Save
-            </Button>
-            <Button variant="outlined" onClick={closeDialog}>
-              Close
-            </Button>
-          </div>
-        </Box>
-      </Modal>
     </>
   );
 }
